app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary   
python-dotenv
psycopg2-binary
